Created Reaction scope and reactor events.
- Now there is an ability to fire off custom events through the filter, events
  that are not associated with with an HTTP request.
- Created Reaction scope for objects that last the duration of a reaction, which
  can be either an HTTP event, or a custom event.
- Created application scope for objects that last the duration of the
  application.
- Added the ability to bind to reaction events to the connector.
- Created a startup event for application initialization.

+++ src/test/java/com/goodworkalan/paste/BinderTest.java
@@ -51,7 +51,7 @@ public class BinderTest
         assertEquals(found.size(), 1);
         assertEquals((int) Tuple.get1(found.get(0)), 0);
         assertTrue(Tuple.get2(found.get(0)) instanceof Forward);
-        found =  bindings.test().put(BindKey.CONTROLLER_CLASS, (Integer) 1).get();
+        found =  bindings.test().put(BindKey.CONTROLLER_CLASS, 1).get();
         assertEquals(found.size(), 0);
     }
 }
